Remove autocast

+++ lumina_next_t2i_mini/train_vae_swap_layers.py
@@ -166,7 +166,7 @@ def main(args):
     logger.info(f"Creating model {dataset_path}")
     model = NextDiT(patch_size=2, dim=2304, n_layers=24, n_heads=32, n_kv_heads=8, cap_feat_dim=cap_feat_dim)
     model.to(device)
-    model.half()
+    # model.half()
     torch.cuda.synchronize()
 
     logger.info(f"Loading model model {dataset_path}")
